requirements.txt
praat-parselmouth
pandas
wheel
scipy
# my-voice-analysis